% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coale_demeny_ex}
\alias{coale_demeny_ex}
\title{Coale Demeny life expectancy}
\format{
[\code{data.table()}] columns:
\describe{
\item{cd_region}{[\code{character()}]\cr "north", "south", "west", or "east"}
\item{sex}{[\code{character()}]\cr "male", "female", or "all}
\item{index}{[\code{numeric()}]\cr Life table level, based on life expectancy at age 10}
\item{age_start}{[\code{numeric()}]\cr Lower bound of age interval}
\item{ex}{[\code{numeric()}]\cr Life expectancy}
\item{ratio_30d10_20d40}{[\code{numeric()}]\cr Ratio 30d10 to 20d40}
}
}
\usage{
coale_demeny_ex
}
\description{
Life expectancy and ratio of 30d10 to 20d40 from Coale
Demeny life tables, to be used for approximation of open age interval
life expectancy, which is an input to SEG DDM estimation. Life tables
are downloaded using the \code{demogR} package.
}
\examples{
coale_demeny_ex
}
\keyword{datasets}
