% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths_per_year.R
\name{deaths_per_year}
\alias{deaths_per_year}
\title{Estimate average annual deaths between censuses}
\usage{
deaths_per_year(
  years_deaths,
  input_deaths,
  years_census,
  step_size = 0.001,
  rule = 2,
  ...
)
}
\arguments{
\item{years_deaths}{[\code{numeric()}]\cr
Which years reported deaths correspond to.}

\item{input_deaths}{[\code{numeric()}]\cr
How many deaths were recorded for each of the years in \code{years_deaths}.}

\item{years_census}{[\code{numeric(2)}]\cr
Vector of two numerics, where each number is a census year, and together
they form the bounds of the intercensal period.}

\item{step_size}{[\code{numeric()}]\cr
Interpolate annual deaths at intervals of \code{step_size}. Default is 0.001
and annual deaths at intervals of 0.001 years will be approximated with
interpolation.}

\item{rule}{[\code{numeric()}]\cr
Rule to be passed to \code{stats::approxfun} for extrapolation.
Default is '2' and constant extrapolation from the extremes is used.}

\item{...}{Other arguments to be passed to \code{stats::approxfun}.}
}
\value{
[\code{numeric(1)}]\cr
Average annual deaths in the period.
}
\description{
Use annual death counts for all or some years between
two censuses to approximate average annual deaths in the
intercensal period. Used to prep input data for death distribution methods.
}
\details{
If only one year of death counts is available, that count is returned as the
average annual death count. If multiple years are available, linear
interpolation is used to approximate annual death counts at all points
in the intercensal period, and then a simple average is taken. Constant
extrapolation is used at the bounds (\code{stats::approxfun} rule = 2) by
default.
}
\examples{
annual_deaths <- deaths_per_year(
  years_deaths = c(2000, 2005, 2006),
  input_deaths = c(100, 110, 114),
  years_census = c(2000, 2010)
)

}
