% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggb}
\alias{ggb}
\title{General Growth Balance (GGB) DDM}
\usage{
ggb(
  dt,
  age_trim_lower = 15,
  age_trim_upper = 65,
  id_cols = "age_start",
  migration = F,
  input_deaths_annual = T,
  input_migrants_annual = T,
  method_growth_rate = "a"
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Input data, must include columns 'pop1', 'pop2', 'deaths', 'age_start',
'date1', 'date2', and all \code{id_cols}. Optional column 'migrants' to be
included if \code{migration = T}.}

\item{age_trim_lower}{[\code{numeric(1)}]\cr
age-trim: lower bound of 'age_start' included (inclusive)}

\item{age_trim_upper}{[\code{numeric(1)}]\cr
age-trim: upper bound of 'age_start' included (exclusive)}

\item{id_cols}{[\code{character()}]\cr
Column names that uniquely identify rows of \code{dt}.
One column must be 'age_start'. When 'age_start' is removed, these
columns define the groups to calculate completeness for.}

\item{migration}{[\code{logical(1)}]\cr
Whether a 'migrants' column should be used. If migration
is available and of reasonable quality, it improves the DDM estimation.
However, these data are not often available.}

\item{input_deaths_annual}{[\code{logical(1)}]\cr
Whether input deaths are formatted as annual average recorded per year.
Alternative is recorded deaths for the entire period between census 1
and census 2. Default TRUE and assumes deaths are annual. If FALSE,
deaths will be divided by the decimal number of years between 'date1' and
'date2'.}

\item{input_migrants_annual}{[\code{logical(1)}]\cr
Whether input migrants are formatted as annual average recorded per year.
Alternative is recorded net migrants for the entire period between census 1
and census 2. Default TRUE and assumes net migrants are annual. If FALSE,
migrants will be divided by the decimal number of years between 'date1' and
'date2'.}

\item{method_growth_rate}{[\code{character(1)}]\cr
Method to use for computation of growth rate. One of:
\itemize{
\item "a": Equation 8c from Hill, You, Choi (2009).
\eqn{r(a+) = (1 / t) * ln(N2(a+) / N1(a+))}.
\item "b": Equation from IUSSP website and Tim Riffe DDM package.
\eqn{r(a+) = (1 / t) * (N2(a+) - N1(a+)) / (exp((log(N1(a+)) +
       log(N2(a+))) / 2))}.
}}
}
\value{
[\code{list(2)}]\cr
\itemize{
\item [\code{data.table()}] Input \code{dt} returned with additional variables
computed in GGB estimation, such as birthdays age a, population age
a plus, entry rate, growth rate, and partial death rate.
\item [\code{data.table()}] Slope, intercept, and estimated completeness by
unique combination of \code{id_cols}.
}
}
\description{
Estimate completeness of death registration using two
censuses with age-specific population, and average annual age-specific
deaths recorded between the censuses, following the General Growth Balance
method.
}
\details{
Columns of \code{dt}:
\itemize{
\item 'pop1': age-specific population recorded at census 1
\item 'pop2': age-specific population recorded at census 2
\item 'deaths': average annual deaths recorded between census 1 and census 2,
or, if 'input_deaths_annual' is FALSE, total deaths recorded between
census 1 and census 2
\item 'age_start': lower bound of age interval
\item 'date1': date of first census, format yyyy-mm-dd
\item 'date2': date of second census, format yyyy-mm-dd
\item 'migrants' (optional): average annual net migrants between census 1 and
census 2, or, if 'input_migrants_annual' is FALSE, total net migrants
recorded between census 1 and census 2
}
}
\examples{
library(data.table)
dt <- copy(zaf_2001_2007)
id_cols <- c("location", "sex", "age_start")
age_trim_lower = 5
age_trim_upper = 85
results <- ggb(
  dt,
  age_trim_lower = age_trim_lower,
  age_trim_upper = age_trim_upper,
  id_cols = id_cols,
  migration = TRUE,
  input_deaths_annual = FALSE,
  input_migrants_annual = FALSE
)

}
\references{
Methods are based on the following sources:
\itemize{
\item http://demographicestimation.iussp.org/content/generalized-growth-balance-method
\item Hill K, You D, Choi Y. Death distribution methods for estimating adult
mortality: sensitivity analysis with simulated data errors. Demographic
Research. 2009 Jul 1;21:235-54.
(https://www.demographic-research.org/volumes/vol21/9/default.htm)
}
}
