% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{plot_ggb}
\alias{plot_ggb}
\title{Plot GGB regression}
\usage{
plot_ggb(results, id_cols_subset)
}
\arguments{
\item{results}{[\code{list(2)}]\cr
Output from \code{ggb} function}

\item{id_cols_subset}{[\code{list()}]\cr
Named list where names are \code{id_cols} and each entry has the value(s) of
the ID columns to be plotted.
Ex: id_cols_subset = list(location = "South Africa", sex = "male",
age_start = seq(5, 80, 5)).}
}
\value{
ggplot2 plot object to display the GGB regression inputs and
linear fit.
}
\description{
Return ggplot2 object with plot for GGB regression using
outputs from \code{ggb} function.
}
