% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zaf_2001_2007}
\alias{zaf_2001_2007}
\title{South Africa male input data for 2001-2007 period}
\format{
[\code{data.table()}] columns:
\describe{
\item{location}{[\code{character()}]\cr "South Africa"}
\item{pop1}{[\code{numeric()}]\cr Age-specific population from census 1}
\item{pop2}{[\code{numeric()}]\cr Age-specific population from census 2}
\item{deaths}{[\code{numeric()}]\cr Age-specific deaths recorded in intercensal period}
\item{age}{[\code{integer()}]\cr Lower bound of 5-year age group}
\item{sex}{[\code{integer()}]\cr "male"}
\item{date1}{[\code{numeric()}]\cr Date of first census}
\item{date2}{[\code{numeric()}]\cr Date of second census}
}
}
\usage{
zaf_2001_2007
}
\description{
Input deaths and population counts for the 2001-2007 period
for males in South Africa. Comes from example here:
http://demographicestimation.iussp.org/content/generalized-growth-balance-method.
}
\examples{
zaf_2001_2007
}
\keyword{datasets}
