% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb.R
\name{ggb}
\alias{ggb}
\title{General Growth Balance (GGB) DDM}
\usage{
ggb(
  dt,
  age_trim_lower = 15,
  age_trim_upper = 65,
  id_cols = "age_start",
  migration = F
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Input data, must include columns 'pop1', 'pop2', 'deaths', 'age_start',
'date1', 'date2', and all \code{id_cols}. Optional column 'migrants' to be
included if \code{migration = T}.}

\item{age_trim_lower}{[\code{numeric(1)}]\cr
age-trim lower bound}

\item{age_trim_upper}{[\code{numeric(1)}]\cr
age trim upper bound}

\item{id_cols}{[\code{character()}]\cr
Column names that uniquely identify rows of \code{dt}.
One column must be 'age_start'. When 'age_start' is removed, these
columns define the groups to calculate completeness for.}

\item{migration}{[\code{logical(1)}]\cr
Whether a 'migrants' column should be used. Then, the balancing
equation is growth = entries - deaths + net migration. If migration
is available and of reasonable quality, it improves the GGB estimation.
However, these data are not often available.}
}
\value{
[\code{list(2)}]\cr
\itemize{
\item [\code{data.table()}] Input \code{dt} returned with additional variables
computed in GGB estimation, such as birthdays age a, population age
a plus, entry rate, growth rate, and partial death rate.
\item [\code{data.table()}] Slope, intercept, and estimated completeness by
unique combination of \code{id_cols}.
}
}
\description{
Estimate completeness of death registration using two
censuses with age-specific population, and average annual age-specific
deaths recorded between the censuses, following the General Growth Balance
method.
}
\details{
Columns of \code{dt}:
\itemize{
\item 'pop1': age-specific population recorded at census 1
\item 'pop2': age-specific population recorded at census 2
\item 'deaths': average annual deaths recorded between census 1 and census 2
\item 'age_start': lower bound of age interval
\item 'date1': date of first census, format yyyy-mm-dd
\item 'date2': date of second census, format yyyy-mm-dd
\item 'migrants' (optional): average annual net migrants between census 1 and
census 2
}
}
\examples{
library(data.table)
dt <- copy(zaf_2001_2007)
setnames(dt, "age", "age_start")
id_cols <- c("location", "sex", "age_start")
age_trim_lower = 5
age_trim_upper = 85
# convert from total to average annual deaths
dt[, deaths := deaths / 5.353425]
results <- ggb(
  dt,
  age_trim_lower = age_trim_lower,
  age_trim_upper = age_trim_upper,
  id_cols = id_cols,
  migration = FALSE
)

}
\references{
Methods are based on the following sources:
\itemize{
\item http://demographicestimation.iussp.org/content/generalized-growth-balance-method
\item Hill K, You D, Choi Y. Death distribution methods for estimating adult
mortality: sensitivity analysis with simulated data errors. Demographic
Research. 2009 Jul 1;21:235-54.
(https://www.demographic-research.org/volumes/vol21/9/default.htm)
}
}
